use-case-test-a2a/requirements.txt
python-dotenv
uvicorn
a2a-sdk[http-server]>=0.3.0
openai-agents
litellm
